requirements.txt
Flask==2.3.2
numpy
opencv-python
scikit-image
pillow
